ensure feature is destroyed to release memory
I think feature.Destroy() will cause g to be destroyed.

The seg fault mentioned in the deleted comment happens because when defer g.Destroy() and defer feature.Destroy() are included, whichever happens second tries to destroy the geometry object that was already removed by the other call.

When only g.Destroy() is called, g is destroyed, but the rest of the feature remains causing the memory leak.